# openapi.yaml
openapi: 3.0.3
info:
  title: Gaardrail API
  version: 1.0.0
  description: API de controle de fluxo baseada em PID para regulação de mensagens em filas
servers:
  - url: http://localhost:8080
    description: Local
paths:
...
  /controller/type:
    put:
      summary: Troca o controlador ativo em runtime (pid ↔ step)
      operationId: switchController
      tags: [PID]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwitchControllerRequest"
            example:
              type: step
      responses:
        "204":
          description: Controlador trocado com sucesso
        "400":
          description: JSON inválido ou campo type ausente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Tipo de controlador desconhecido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    SwitchControllerRequest:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [pid, step]
          description: Tipo do controlador a ativar
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string